try with int, hmm

+++ test.py
@@ -5,7 +5,7 @@ servos = ServoDriver()
 try:
     while True:
         raw_value = input("Enter speed: ")
-        servo_value = float(raw_value)
+        servo_value = int(raw_value)
         servos.set_throttle(servo_value)
 except KeyboardInterrupt:
     pass
